[ui-components] refactor: moved infoicon/radiogroup/checkbox (#2619)
