+++ bundle-analyzer/bundlesSizesBaseline.json
@@ -13,7 +13,7 @@
   "wix-rich-content-plugin-image": 258,
   "wix-rich-content-plugin-indent": 240,
   "wix-rich-content-plugin-line-spacing": 243,
-  "wix-rich-content-plugin-link-preview": 258,
+  "wix-rich-content-plugin-link-preview": 257,
   "wix-rich-content-plugin-link": 243,
   "wix-rich-content-plugin-map": 369,
   "wix-rich-content-plugin-mentions": 243,
@@ -31,14 +31,14 @@
   "ricos-viewer": 202,
   "ricos-viewer-no-modal": 202,
   "ricos-viewer-with-plugins": 275,
-  "rce": 759,
-  "rce-with-plugins": 985,
-  "ricos-editor": 817,
-  "ricos-editor-no-modal": 817,
-  "ricos-editor-with-plugins": 1043,
+  "rce": 786,
+  "rce-with-plugins": 983,
+  "ricos-editor": 840,
+  "ricos-editor-no-modal": 840,
+  "ricos-editor-with-plugins": 1037,
   "ricos-common": 20,
   "ricos-content": 55,
   "rich-content-schema": 57,
   "converters": 316,
-  "editor-common": 448
+  "editor-common": 474
 }
