+++ packages/editor-common/web/src/Icons/index.ts
@@ -4,7 +4,6 @@ export { default as AlignTextCenterIcon } from './AlignTextCenterIcon';
 export { default as AlignJustifyIcon } from './AlignJustifyIcon';
 export { default as AlignLeftIcon } from './AlignLeftIcon';
 export { default as AlignRightIcon } from './AlignRightIcon';
-export { default as CheckIcon } from './CheckIcon';
 export { default as DropdownArrowIcon } from './DropdownArrowIcon';
 export { default as ErrorIcon } from './ErrorIcon';
 export { default as LinkIcon } from './LinkIcon';
