+++ packages/ui-components/web/src/Components/RadioGroupVertical.jsx
@@ -1,9 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { RadioGroup, InfoIcon } from 'wix-rich-content-editor-common';
+import RadioGroup from './RadioGroup';
+import InfoIcon from './InfoIcon';
 import { mergeStyles } from 'wix-rich-content-common';
 import styles from '../../statics/styles/radio-group-vertical.scss';
-import generalstyles from 'wix-rich-content-editor-common/dist/statics/styles/general.scss';
+import infoIconStyles from '../../statics/styles/info-icon.scss';
 
 class RadioGroupVertical extends Component {
   constructor(props) {
@@ -19,7 +20,7 @@ class RadioGroupVertical extends Component {
 
     return (
       <div>
-        <div className={generalstyles.infoContainer}>
+        <div className={infoIconStyles.infoContainer}>
           {label ? (
             <span id={`${this.id}_label`} className={styles.radioGroupVertical_title}>
               {label}
