+++ packages/ui-components/web/statics/styles/checkbox.scss
@@ -61,3 +61,7 @@
 .checkbox_label {
   margin-left: 12px;
 }
+
+// *:focus,
+.focused {
+}
