+++ packages/ui-components/web/statics/styles/file-input.scss
@@ -1,0 +1,12 @@
+// *:focus,
+.focused {
+}
+
+.visuallyHidden {
+  width: 0.1px;
+  height: 0.1px;
+  overflow: hidden;
+  position: absolute;
+  opacity: 0;
+  z-index: -1;
+}
